requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.72.1
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.1
attrs==25.4.0
Authlib==1.6.5
beartype==0.22.5
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
cloudpickle==3.1.2
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
cyclopts==4.2.3
dataclasses-json==0.6.7
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
docx==0.2.4
email-validator==2.3.0
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.13.2
fakeredis==2.33.0
fastapi==0.128.0
fastmcp==2.14.0
filelock==3.20.0
firecrawl==4.12.0
Flask==3.1.2
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
ipython==9.9.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.6.0
langchain==1.0.5
langchain-anthropic==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-mcp-adapters==0.1.13
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.42
looseversion==1.3.0
lupa==2.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mcp==1.25.0
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.4
openai==2.8.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathable==0.4.4
pathlib==1.0.1
pathvalidate==3.3.1
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prometheus_client==0.24.0
prompt_toolkit==3.0.52
propcache==0.4.1
prov==2.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.6
pydot==4.0.1
pyee==11.1.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.2.5
pypdf==6.2.0
pypdfium2==5.1.0
pyperclip==1.11.0
pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-louvain==0.16
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.4.0
redis==7.1.0
referencing==0.36.2
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
scipy==1.16.3
setuptools==78.1.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
tavily-python==0.7.17
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uv==0.9.22
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
Werkzeug==3.1.5
wheel==0.45.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
